Add Feature submit form with data to backend and image to S3 bucket
