+++ src/Components/UploadImage.js
@@ -1,25 +1,9 @@
 import { TextField } from '@mui/material';
 import React, { useState } from 'react';
 import PropTypes from 'prop-types';
-import AWS from 'aws-sdk';
-
-const S3_BUCKET = 'bee-bucket-microverse';
-const REGION = 'us-east-1';
-
-AWS.config.update({
-  accessKeyId: process.env.REACT_APP_AWS_BUCKET_ACCESS_KEY,
-  secretAccessKey: process.env.REACT_APP_AWS_BUCKET_SECRET_KEY,
-});
-
-const myBucket = new AWS.S3({
-  params: { Bucket: S3_BUCKET },
-  region: REGION,
-});
 
 const UploadImage = (props) => {
-  console.log(process.env.REACT_APP_AWS_BUCKET_ACCESS_KEY);
-  const { selectedFile, setSelectedFile } = props;
-  const [progress, setProgress] = useState(0);
+  const { setSelectedFile, progress } = props;
   const [disableSubmit, setDisableSubmit] = useState(true);
 
   const handleFileInput = (e) => {
@@ -35,22 +19,6 @@ const UploadImage = (props) => {
     }
   };
 
-  const uploadFile = (file) => {
-    const params = {
-      ACL: 'public-read',
-      Body: file,
-      Bucket: S3_BUCKET,
-      Key: file.name,
-    };
-
-    myBucket
-      .putObject(params)
-      .on('httpUploadProgress', (evt) => {
-        setProgress(Math.round((evt.loaded / evt.total) * 100));
-      })
-      .send((err) => err);
-  };
-
   return (
     <div>
       <TextField
@@ -63,17 +31,13 @@ const UploadImage = (props) => {
       />
       <button
         type="button"
-        onClick={() => uploadFile(selectedFile)}
+        // onClick={() => uploadFile(selectedFile)}
         disabled={disableSubmit}
       >
         {' '}
         Upload Image
       </button>
-      <div>
-        File Upload Progress is
-        {' '}
-        {progress}
-      </div>
+      <div>File Upload Progress is {progress}</div>
     </div>
   );
 };
@@ -86,6 +50,7 @@ UploadImage.propTypes = {
     size: PropTypes.number.isRequired,
   }),
   setSelectedFile: PropTypes.func.isRequired,
+  progress: PropTypes.number.isRequired,
 };
 
 export default UploadImage;
