+++ src/Pages/AddBeePage.js
@@ -1,4 +1,5 @@
 import { Button, TextField } from '@mui/material';
+import AWS from 'aws-sdk';
 import LibraryAddIcon from '@mui/icons-material/LibraryAdd';
 import React, { useState } from 'react';
 import UploadImage from '../Components/UploadImage';
@@ -6,20 +7,47 @@ import '../Styles/BeeForm.scss';
 
 const AddBeePage = () => {
   const [selectedFile, setSelectedFile] = useState(null);
+  const [progress, setProgress] = useState(0);
+
+  const S3_BUCKET = 'bee-bucket-microverse';
+  const REGION = 'us-east-1';
+
+  AWS.config.update({
+    accessKeyId: process.env.REACT_APP_AWS_BUCKET_ACCESS_KEY,
+    secretAccessKey: process.env.REACT_APP_AWS_BUCKET_SECRET_KEY,
+  });
+
+  const myBucket = new AWS.S3({
+    params: { Bucket: S3_BUCKET },
+    region: REGION,
+  });
+
+  const uploadFile = (file) => {
+    const params = {
+      ACL: 'public-read',
+      Body: file,
+      Bucket: S3_BUCKET,
+      Key: file.name,
+    };
+
+    myBucket
+      .putObject(params)
+      .on('httpUploadProgress', (evt) => {
+        setProgress(Math.round((evt.loaded / evt.total) * 100));
+      })
+      .send((err) => err);
+  };
 
   const handleSubmit = (e) => {
     e.preventDefault();
+    uploadFile(selectedFile);
   };
 
   return (
     <form className="bee-form" onSubmit={handleSubmit}>
       <h1>Add Your own Bee</h1>
       <div>
-        <TextField
-          required
-          id="name-form"
-          label="Name"
-        />
+        <TextField required id="name-form" label="Name" />
       </div>
       <div>
         <TextField
@@ -31,10 +59,7 @@ const AddBeePage = () => {
         />
       </div>
       <div>
-        <UploadImage
-          selectedFile={selectedFile}
-          setSelectedFile={setSelectedFile}
-        />
+        <UploadImage setSelectedFile={setSelectedFile} progress={progress} />
       </div>
       <Button
         variant="contained"
